Converted to Typescript
