+++ package.json
@@ -6,28 +6,6 @@
   "private": true,
   "homepage": "https://react-native-elements.github.io/react-native-elements-app",
   "main": "src/index.native.js",
-  "dependencies": {
-    "@babel/polyfill": "^7.2.5",
-    "@expo/vector-icons": "^9.0.0",
-    "expo": "^32.0.0",
-    "expo-linear-gradient": "^3.0.0",
-    "gh-pages": "^2.0.1",
-    "lodash": "^4.17.4",
-    "prop-types": "^15.7.2",
-    "react": "16.5.0",
-    "react-art": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
-    "react-native-elements": "^1.1.0",
-    "react-native-gesture-handler": "1.0.17",
-    "react-native-gesture-handler-web": "npm:react-native-gesture-handler@1.1.0",
-    "react-native-ratings": "^6.3.0",
-    "react-native-touchable-scale": "2.0.0",
-    "react-native-vector-icons": "^6.3.0",
-    "react-native-web": "^0.10.0",
-    "react-navigation": "^3.5.1",
-    "react-scripts": "^2.1.5"
-  },
   "scripts": {
     "web": "react-app-rewired start",
     "native": "expo start",
@@ -38,14 +16,53 @@
     "eject:web": "react-scripts eject",
     "eject:native": "expo eject",
     "prettify": "prettier --write 'src/**/*.js'",
-    "clean-install": "rm -rf node_modules && npm cache clean --force && watchman watch-del-all && yarn"
+    "clean": "rm -rf node_modules && npm cache clean --force && watchman watch-del-all && yarn"
+  },
+  "dependencies": {
+    "@expo/vector-icons": "^9.0.0",
+    "expo": "^32.0.0",
+    "expo-linear-gradient": "^3.0.0",
+    "gh-pages": "^2.0.1",
+    "lodash": "^4.17.4",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-art": "^16.8.3",
+    "react-dom": "^16.8.6",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "react-native-elements": "^1.1.0",
+    "react-native-gesture-handler": "1.0.17",
+    "react-native-gesture-handler-web": "npm:react-native-gesture-handler@1.1.0",
+    "react-native-ratings": "^6.3.0",
+    "react-native-touchable-scale": "2.0.0",
+    "react-native-vector-icons": "^6.3.0",
+    "react-native-web": "^0.10.0",
+    "react-navigation": "^3.5.1",
+    "react-scripts": "^3.0.1",
+    "tslint": "^5.18.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "3.4.5"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.3.4",
-    "babel-preset-expo": "^5.0.0",
-    "customize-cra": "^0.2.12",
-    "prettier": "^1.15.3",
-    "react-app-rewired": "^2.1.0"
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/polyfill": "^7.4.4",
+    "@babel/preset-env": "^7.5.5",
+    "@types/expo": "^32.0.13",
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.6.8",
+    "@types/react": "^16.8.23",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-native": "^0.60.2",
+    "@types/react-test-renderer": "^16.8.3",
+    "babel-preset-expo": "^6.0.0",
+    "customize-cra": "^0.4.1",
+    "prettier": "^1.18.2",
+    "react-app-rewired": "^2.1.3"
+  },
+  "engineStrict": true,
+  "engines": {
+    "node": ">=10.16.0 <12.0.0",
+    "npm": ">9999.0.0",
+    "yarn": ">=1.9.4"
   },
   "browserslist": [
     ">0.2%",
