--- src/components/AppLoading.js
@@ -1,1 +1,0 @@
-export {AppLoading as default} from "expo";
