+++ src/components/AppLoading.tsx
@@ -1,0 +1,1 @@
+export {AppLoading} from "expo";
