+++ src/config/fonts.ts
