+++ src/config/socialColors.ts
