+++ src/helpers/AssetsCaching.web.ts
@@ -1,20 +1,25 @@
 export const cacheAssets = () => Promise.resolve(null);
 
-let cachedFonts = {};
-const cacheFont = (name, link) => {
+
+
+interface fontsType {
+  [name: string]: string
+}
+
+let cachedFonts: fontsType = {};
+const cacheFont = (name: string, link: string) => {
   const styleBody = `@font-face { src: url(${link}); font-family: ${name}; }`;
   const style = document.createElement("style");
   style.type = "text/css";
-  if (style.styleSheet) {
-    style.styleSheet.cssText = styleBody;
-  } else {
-    style.appendChild(document.createTextNode(styleBody));
-  }
+  // @ts-ignore
+  if (style.styleSheet) style.styleSheet.cssText = styleBody;
+  else style.appendChild(document.createTextNode(styleBody));
+
   document.head.appendChild(style);
   cachedFonts[name] = link;
 };
 
-export const cacheFonts = fonts => {
+export const cacheFonts = (fonts: fontsType) => {
   let jobs = [];
   for(let fontName in fonts) {
     if(!cachedFonts[fontName])
