+++ src/index.web.tsx
@@ -1,5 +1,3 @@
-import "@babel/polyfill";
-
 import React from "react";
 import App from "./App";
 import ReactDOM from "react-dom";
