+++ tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["tslint:recommended", "tslint-config-prettier"],
+  "rules": {
+    "interface-name": false
+  }
+}
